major work on cleaning the implementation

+++ fe_cmov.c
@@ -7,7 +7,7 @@ replace (f,g) with (f,g) if b == 0.
 Preconditions: b in {0,1}.
 */
 
-void fe_cmov(fe f,const fe g,unsigned int b)
+void fe_cmov(fe f, const fe g, unsigned int b)
 {
   int32_t f0 = f[0];
   int32_t f1 = f[1];
@@ -39,7 +39,7 @@ void fe_cmov(fe f,const fe g,unsigned int b)
   int32_t x7 = f7 ^ g7;
   int32_t x8 = f8 ^ g8;
   int32_t x9 = f9 ^ g9;
-  b = -b;
+  b = (unsigned int) (- (int) b); /* silence warning */
   x0 &= b;
   x1 &= b;
   x2 &= b;
